pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langvio"
version = "0.0.1"
description = "Connect language models to vision models for natural language visual analysis"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Mughees Mehdi", email = "mugheesmehdi@gmail.com" }
]

keywords = ["computer-vision", "llm", "ai", "yolo", "langchain", "object-detection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
    "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.8,<3.13"
dependencies = [
    "torch>=1.9.0",
    "ultralytics>=8.0.0",
    "opencv-python>=4.5.0",
    "numpy>=1.20.0",
    "pillow>=8.0.0",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.10",
    "pyyaml>=6.0",
    "python-dotenv>=0.19.0",
    "tqdm>=4.64.0",
]

[project.optional-dependencies]
openai = [
    "openai>=1.0.0",
    "langchain-openai>=0.0.1",
]
google = [
    "google-generativeai>=0.3.0,<1.0.0",
    "langchain-google-genai>=2.0.0",
]

all-llm = [
    "langvio[openai,google]",
]
webapp = [
    "flask>=2.0.0",
    "werkzeug>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/MugheesMehdi07/langvio"
Documentation = "https://github.com/MugheesMehdi07/langvio#readme"
Repository = "https://github.com/MugheesMehdi07/langvio"
Issues = "https://github.com/MugheesMehdi07/langvio/issues"
Changelog = "https://github.com/MugheesMehdi07/langvio/blob/main/CHANGELOG.md"

[project.scripts]
langvio = "langvio.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["langvio*"]
exclude = ["tests*", "examples*", "webapp*", "docs*"]

[tool.setuptools.package-data]
langvio = ["*.yaml", "*.yml", "*.json"]


